utils: Move Git ref helpers
The Git utility layer has named modules for object IO, repository identity,
index plumbing, and worktree operations, while ref transactions still share
the command execution wrapper.

The project lacks a named boundary for update-ref transactions.
Batch-source and undo bookkeeping code must import ref mutation from the
broad command module.

This commit addresses that boundary gap by moving update_git_refs and the
private ref existence probe into utils.git_refs, then updating source
callers to import that module directly.

The next commit will add architecture coverage for the Git ref helper
boundary.
