+++ src/git_stage_batch/data/batch_sources.py
@@ -20,6 +20,8 @@ from ..i18n import _
 from ..utils.file_io import read_file_paths_file, read_text_file_contents, write_text_file_contents
 from ..utils.git import (
     run_git_command,
+)
+from ..utils.git_refs import (
     update_git_refs,
 )
 from ..utils.git_index import (
