+++ src/git_stage_batch/data/undo_restore.py
@@ -20,6 +20,8 @@ from ..i18n import _
 from ..utils.file_io import read_file_paths_file
 from ..utils.git import (
     run_git_command,
+)
+from ..utils.git_refs import (
     update_git_refs,
 )
 from ..utils.git_worktree import git_checkout_detached
